pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["amadeo <risefenixpuryfire@gmail.com>"]
readme = "README.md"



[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.38"
greenlet = "^3.1.1"
asyncpg = "^0.30.0"
selenium = "^4.29.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.0"
webdriver-manager = "^4.0.2"
aiogram = "^3.18.0"
aioschedule = "^0.5.2"
schedule = "^1.2.2"
fastapi = "^0.115.9"
uvicorn = "^0.34.0"
chromedriver-binary = "^135.0.7039.0.0"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
